feat(reader): consolidate CbzReaderViewModel into shared commonMain (#953)
Move the full 896-line CbzReaderViewModel from :app into :feature:reader/commonMain
so Android and iOS run one shared class (replacing the 85-line iOS stub). The VM is
now a plain androidx.lifecycle.ViewModel: the nav itemId is a constructor param
(Android Koin reads it from SavedStateHandle at the call site), dispatchers come from
DispatcherProvider, and the position locator JSON is built/parsed with
kotlinx.serialization instead of org.json. CbzReaderState moves with it.

Android: Koin points at the shared class; CbzReaderScreen imports the new package.
iOS: IosCbzReaderScreen hosts the shared VM via a Koin factory keyed by itemId and
renders CbzReaderState; new iOS backends (IosCbzRepository over the existing
downloader/Komga streaming source, plus no-op stores for sync/wakelock/formatting/
panel-mask/report/appearance/developer-options, following the IosNoOp* pattern) satisfy
the VM's dependency graph. Real panel detection already runs on iOS via the
PanelEngine added earlier.

Tests: the pure-helper tests (CbzArchiveSwapTest, ComicFormattingOverridesMergeTest,
SmartSeamHoldTest) move to :feature:reader/commonTest and now run on JVM + iOS. The
throwaway iOS stub VM and its prototype-API test are retired (the stub's panel-nav API
no longer exists in that form; the shared VM's behaviour supersedes it).

All green: feature:reader jvm+ios tests, app compile + cbz unit tests, shared iosArm64
+ iosSimulatorArm64 compile, checkNoAndroidImports, ktlint (shared/core:data iosMain).

Progresses #953.

Removed-test: initial page is 0
Removed-test: gotoPage updates currentPage
Removed-test: gotoPage clamps to valid range
Removed-test: gotoPage fires onPositionChanged
Removed-test: gotoPage does not fire onPositionChanged when page unchanged
Removed-test: panel index starts at -1
Removed-test: nextPanel with fallback panels stays at -1
Removed-test: enterPanelView with real panels sets panelIndex to 0
Removed-test: enterPanelView with fallback panels does not activate panel view
Removed-test: nextPanel advances panel index
Removed-test: nextPanel at last panel goes to next page
Removed-test: prevPanel at first panel goes to prev page last panel
Removed-test: exitPanelView resets panelIndex to -1
Removed-test: pageCount matches imageSource
Removed-test: returns null when actualPageCount is zero (no clamping, keep streaming count)
