--- app/src/main/kotlin/com/riffle/app/feature/reader/cbz/CbzReaderViewModel.kt
@@ -1,797 +1,0 @@
-package com.riffle.app.feature.reader.cbz
-
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.viewModelScope
-import com.riffle.app.feature.reader.ReaderStateHolder
-import com.riffle.app.feature.reader.RailSegment
-import com.riffle.feature.reader.VolumeNavEvent
-import com.riffle.app.feature.reader.VolumeNavigationController
-import com.riffle.app.feature.reader.buildCbzRailSegments
-import com.riffle.app.feature.reader.cbzRailCursorPosition
-import com.riffle.app.feature.reader.controllers.VolumeKeyDispatcher
-import com.riffle.app.feature.reader.findActiveCbzSegmentIndex
-import com.riffle.core.domain.CbzLocalSource
-import com.riffle.core.domain.CbzOpenResult
-import com.riffle.core.domain.CbzRepository
-import com.riffle.core.domain.LibraryObserver
-import com.riffle.core.domain.ProgressSyncController
-import com.riffle.core.domain.ReadingSessionRepository
-import com.riffle.core.domain.ReaderTheme
-import com.riffle.core.domain.WakeLockPreferencesStore
-import com.riffle.core.domain.appearance.AppearanceCoordinator
-import com.riffle.core.models.LibraryItem
-import com.riffle.core.models.SessionPayload
-import com.riffle.core.domain.comic.BookComicFormattingOverrides
-import com.riffle.core.domain.comic.BookComicFormattingPreferencesStore
-import com.riffle.core.domain.comic.ComicBookmark
-import com.riffle.core.domain.comic.ComicPageSource
-import com.riffle.core.domain.comic.ComicFormattingPreferences
-import com.riffle.core.domain.comic.ComicFormattingPreferencesStore
-import com.riffle.core.domain.comic.PanelOverflowBehavior
-import com.riffle.core.domain.comic.resolveComicBackgroundTheme
-import com.riffle.core.domain.comic.panel.ColorPageDecoder
-import com.riffle.core.domain.comic.panel.PagePanels
-import com.riffle.core.domain.comic.panel.PanelBinaryMask
-import com.riffle.core.domain.comic.panel.PanelEngine
-import com.riffle.core.domain.comic.panel.comicSeamEnergySampler
-import com.riffle.core.domain.comic.panel.PanelMaskService
-import com.riffle.core.domain.comic.panel.PanelOverflowTransform
-import com.riffle.core.domain.comic.panel.PanelRegion
-import com.riffle.core.domain.comic.panel.PanelReportRepository
-import com.riffle.core.domain.comic.panel.PanelSource
-import com.riffle.core.domain.comic.panel.PanelViewPreferencesStore
-import com.riffle.core.domain.developer.DeveloperOptionsRepository
-import com.riffle.core.domain.usecase.UpdateReadingProgress
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import org.json.JSONObject
-
-/**
- * Backs the comic (CBZ) reader. Deliberately smaller than [com.riffle.app.feature.reader.PdfReaderViewModel]:
- * comics don't go through Readium — page images come straight out of a [CbzArchive] — so there's no
- * Publication, no navigator, no formatting session, no annotations, no TOC.
- *
- * v1 scope per ADR 0050; Panel View overlay per ADR 0055 (opt-in per-book toggle that frames one
- * panel at a time using auto-detected regions).
- */
-class CbzReaderViewModel constructor(
-    application: Application,
-    savedStateHandle: SavedStateHandle,
-    private val libraryObserver: LibraryObserver,
-    private val cbzRepository: CbzRepository,
-    private val readingSessionRepository: ReadingSessionRepository,
-    private val updateReadingProgressUseCase: UpdateReadingProgress,
-    private val wakeLockPreferencesStore: WakeLockPreferencesStore,
-    private val volumeNavigationController: VolumeNavigationController,
-    private val volumeKeyDispatcher: VolumeKeyDispatcher,
-    private val readerStateHolder: ReaderStateHolder,
-    private val panelEngine: PanelEngine,
-    private val panelMaskService: PanelMaskService,
-    private val panelViewPreferencesStore: PanelViewPreferencesStore,
-    private val comicFormattingPreferencesStore: ComicFormattingPreferencesStore,
-    private val bookComicFormattingPreferencesStore: BookComicFormattingPreferencesStore,
-    private val developerOptionsRepository: DeveloperOptionsRepository,
-    private val appearanceCoordinator: AppearanceCoordinator,
-    private val colorPageDecoder: ColorPageDecoder,
-    val panelReportRepository: PanelReportRepository,
-) : AndroidViewModel(application) {
-
-    private val itemId: String = checkNotNull(savedStateHandle["itemId"])
-
-    private var currentSource: ComicPageSource? = null
-    private var lastSavedPage: Int = -1
-    private var closeSyncDone: Boolean = false
-    var bookId: String = itemId
-        private set
-    private var panelBook: PanelEngine.Book? = null
-    // Cancel-and-replace the in-flight panel resolve on every page change so a stale resolver
-    // can't overwrite a newer page's PagePanels result.
-    private var panelResolveJob: Job? = null
-    // previousPanel's cross-page landing-index job — tracked separately so it never orphans
-    // the page resolve launched by gotoPage (both must be cancellable on the next page change).
-    private var panelLandingJob: Job? = null
-    // Guard so background prefetches skip store writes after the archive is closed on
-    // onCleared, otherwise a stale coroutine calling into a closed ZipFile persists a bogus
-    // 1x1 Fallback that would poison the next open's cache.
-    @Volatile private var archiveClosed: Boolean = false
-
-    private val syncSession = ProgressSyncController(
-        itemId = itemId,
-        repository = readingSessionRepository,
-        scope = viewModelScope,
-    )
-
-    private val _state = MutableStateFlow<CbzReaderState>(CbzReaderState.Loading)
-    val state: StateFlow<CbzReaderState> = _state
-
-    private val _currentPage = MutableStateFlow(0)
-    val currentPage: StateFlow<Int> = _currentPage
-
-    private val _railSegments = MutableStateFlow<List<RailSegment>>(emptyList())
-    val railSegments: StateFlow<List<RailSegment>> = _railSegments.asStateFlow()
-
-    private val _pageCount = MutableStateFlow(0)
-
-    val activeRailSegmentIndex: StateFlow<Int> = combine(_currentPage, _railSegments) { page, segments ->
-        findActiveCbzSegmentIndex(segments, page).coerceAtLeast(0)
-    }.stateIn(viewModelScope, SharingStarted.Eagerly, 0)
-
-    val railCursorPosition: StateFlow<Float> = combine(_currentPage, _railSegments, _pageCount) { page, segments, pageCount ->
-        val activeIndex = findActiveCbzSegmentIndex(segments, page).coerceAtLeast(0)
-        cbzRailCursorPosition(segments, activeIndex, page, pageCount.coerceAtLeast(1))
-    }.stateIn(viewModelScope, SharingStarted.Eagerly, 0f)
-
-    private val _bookComicOverrides = MutableStateFlow(BookComicFormattingOverrides())
-
-    val effectiveComicFormatting: StateFlow<ComicFormattingPreferences> =
-        combine(comicFormattingPreferencesStore.preferences, _bookComicOverrides) { global, overrides ->
-            overrides.applyTo(global)
-        }.stateIn(viewModelScope, SharingStarted.Eagerly, ComicFormattingPreferences())
-
-    val comicBackgroundTheme: StateFlow<ReaderTheme> =
-        combine(effectiveComicFormatting, appearanceCoordinator.resolved) { formatting, appearance ->
-            formatting.backgroundTheme.resolveComicBackgroundTheme(appearance.readerTheme.toReaderTheme())
-        }.stateIn(viewModelScope, SharingStarted.Eagerly, ReaderTheme.Dark)
-
-    val hasComicOverrides: StateFlow<Boolean> = _bookComicOverrides
-        .map { !it.isEmpty() }
-        .stateIn(viewModelScope, SharingStarted.Eagerly, false)
-
-    /** Panel View toggle for the current book (ADR 0055). Derived from [effectiveComicFormatting]. */
-    val panelViewOn: StateFlow<Boolean> = effectiveComicFormatting
-        .map { it.panelViewOn }
-        .stateIn(viewModelScope, SharingStarted.Eagerly, false)
-
-    private val _viewportSize = MutableStateFlow(0 to 0) // (width, height) in px
-
-    fun setViewportSize(w: Int, h: Int) {
-        _viewportSize.value = w to h
-    }
-
-    /** Low-resolution page colour-image sampler for SMART_SPLIT seam detection. */
-    private val _energySampler = MutableStateFlow<EnergySamplerState>(EnergySamplerState.Resolved(null))
-
-    private val _currentPagePanels = MutableStateFlow<PagePanels?>(null)
-    /**
-     * Panels resolved for the currently-displayed page (row-band ordered). Null while the
-     * detector is still working. When Panel View is off, this is still populated in the
-     * background so re-enabling is instant.
-     */
-    val currentPagePanels: StateFlow<PagePanels?> = _currentPagePanels
-
-    /**
-     * Effective panels after applying the overflow transform (e.g. SPLIT). This is what the
-     * screen renders AND what panel navigation walks — a null emission here keeps the screen
-     * in its loading state (the SMART_SPLIT seam hold below relies on that contract).
-     */
-    val effectivePanels: StateFlow<PagePanels?> = combine(
-        _currentPagePanels,
-        effectiveComicFormatting,
-        _viewportSize,
-        _energySampler,
-    ) { pagePanels, formatting, (vpW, vpH), samplerState ->
-        if (pagePanels == null || pagePanels.isFallback || !formatting.panelViewOn) return@combine pagePanels
-        val overflow = formatting.panelOverflow
-        if ((overflow != PanelOverflowBehavior.SPLIT && overflow != PanelOverflowBehavior.SMART_SPLIT) ||
-            vpW == 0 || vpH == 0) return@combine pagePanels
-        // While the SMART_SPLIT seam decode is still in flight, emitting a centre-split interim
-        // result would render the page and then visibly re-split it when the real seam arrives.
-        // Hold (null → screen keeps its loading state) until the sampler resolves instead.
-        if (shouldHoldForSmartSeam(overflow, samplerState is EnergySamplerState.Pending, pagePanels, vpW, vpH)) {
-            return@combine null
-        }
-        val transformed = PanelOverflowTransform.applyOverflow(
-            pagePanels.panels, pagePanels.imageWidth, pagePanels.imageHeight, vpW, vpH,
-            overflow,
-            energySampler = if (overflow == PanelOverflowBehavior.SMART_SPLIT) {
-                (samplerState as? EnergySamplerState.Resolved)?.sampler
-            } else {
-                null
-            },
-        )
-        pagePanels.copy(panels = transformed)
-    }
-        // applyOverflow's SMART_SPLIT sampler scans pixels — keep it off the main thread,
-        // especially now that the seam hold puts this emission on the first-render path.
-        .flowOn(Dispatchers.Default)
-        .stateIn(viewModelScope, SharingStarted.Eagerly, null)
-
-    private val _currentPanelIndex = MutableStateFlow(0)
-    /** 0-based index into [currentPagePanels]. Advances with next/prev-panel gestures. */
-    val currentPanelIndex: StateFlow<Int> = _currentPanelIndex
-
-    val keepScreenOn: StateFlow<Boolean> = wakeLockPreferencesStore.keepScreenOn
-        .stateIn(viewModelScope, SharingStarted.Eagerly, true)
-
-    val volumeNavEvents: SharedFlow<VolumeNavEvent> = volumeNavigationController.events
-
-    val volumeKeyNavigationEnabled = volumeKeyDispatcher.volumeKeyNavigationEnabled
-        .stateIn(viewModelScope, SharingStarted.Eagerly, true)
-
-    val invertVolumeKeys = volumeKeyDispatcher.invertVolumeKeys
-        .stateIn(viewModelScope, SharingStarted.Eagerly, false)
-
-    val developerModeEnabled: StateFlow<Boolean> = developerOptionsRepository.developerModeEnabled
-        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5_000), false)
-
-    /**
-     * Produces the binarized mask for the report sheet via [PanelMaskService], which uses
-     * the same detection pipeline as the detector so the reported mask is byte-identical to
-     * what was used on-device. See [AndroidPanelMaskServiceImpl] for the invariant details
-     * (issues #783/#784/#786).
-     *
-     * Returns null if the page is a uniform solid or on I/O error.
-     */
-    suspend fun generateMaskPng(pageIndex: Int): Pair<PanelBinaryMask, ByteArray>? =
-        withContext(Dispatchers.IO) {
-            val imageSource = (_state.value as? CbzReaderState.Ready)?.imageSource ?: return@withContext null
-            val rawBytes = runCatching { imageSource.imageBytes(pageIndex) }.getOrNull() ?: return@withContext null
-            panelMaskService.generateMask(pageIndex, rawBytes)
-        }
-
-    init {
-        viewModelScope.launch { openBook() }
-        // Collect server-wins events from the sync cycle and jump to the server's page.
-        viewModelScope.launch {
-            syncSession.serverPositionEvents.collect { serverProgress ->
-                val page = parsePageIndex(serverProgress.ebookLocation) ?: return@collect
-                val ready = _state.value as? CbzReaderState.Ready ?: return@collect
-                val clamped = page.coerceIn(0, ready.pageCount - 1)
-                if (clamped != _currentPage.value) {
-                    _currentPage.value = clamped
-                    _currentPanelIndex.value = 0
-                    _currentPagePanels.value = null
-                    lastSavedPage = clamped
-                    onCurrentPageChanged(clamped)
-                    val progressFraction = if (ready.pageCount > 1) clamped.toFloat() / (ready.pageCount - 1).toFloat() else 1f
-                    viewModelScope.launch { updateReadingProgressUseCase(itemId, progressFraction) }
-                }
-            }
-        }
-    }
-
-    private suspend fun openBook() {
-        // Ensure the global ComicFormattingPreferences DataStore has emitted at least once before
-        // state becomes Ready. Without this, effectiveComicFormatting starts at the default
-        // (panelViewOn=false), causing a pager flash for books with Panel View enabled.
-        comicFormattingPreferencesStore.preferences.first()
-
-        val item = libraryObserver.getItem(itemId)
-        if (item == null) {
-            _state.value = CbzReaderState.Error("Book not found")
-            return
-        }
-        bookId = "${item.sourceId}::${item.id}"
-        // Load per-book Comic formatting overrides (new store).
-        // Also performs a one-time lazy migration from the legacy PanelViewPreferencesStore.
-        val existingOverrides = bookComicFormattingPreferencesStore.overrides(bookId).first()
-        val legacyPanelViewOn = if (existingOverrides.panelViewOn == null) {
-            panelViewPreferencesStore.state(bookId).first().panelViewOn
-        } else null
-
-        if (legacyPanelViewOn == true) {
-            val migrated = existingOverrides.copy(panelViewOn = true)
-            bookComicFormattingPreferencesStore.save(bookId, migrated)
-            _bookComicOverrides.value = migrated
-        } else {
-            _bookComicOverrides.value = existingOverrides
-        }
-        when (val result = cbzRepository.openCbz(item)) {
-            is CbzOpenResult.Success -> loadOpened(
-                result.imageSource, result.pageCount, result.bookmarks, result.lastPosition, item.title,
-            )
-            is CbzOpenResult.Streaming -> loadStreaming(result, item)
-            is CbzOpenResult.NetworkError -> _state.value = CbzReaderState.Error(
-                result.cause.message ?: "Network error"
-            )
-            CbzOpenResult.Offline -> _state.value = CbzReaderState.Error("Book not available offline")
-        }
-    }
-
-    private suspend fun loadOpened(
-        source: ComicPageSource,
-        pageCount: Int,
-        bookmarks: List<ComicBookmark>,
-        lastPosition: String?,
-        title: String,
-    ) {
-        if (pageCount == 0) {
-            _state.value = CbzReaderState.Error("Comic has no pages")
-            source.close()
-            return
-        }
-        currentSource = source
-        _railSegments.value = buildCbzRailSegments(bookmarks, pageCount)
-        _pageCount.value = pageCount
-        val resumeIndex = lastPosition
-            ?.let { parsePageIndex(it) }
-            ?.coerceIn(0, pageCount - 1)
-            ?: 0
-        _currentPage.value = resumeIndex
-        lastSavedPage = resumeIndex
-        // Always start Panel View at the first panel on book open. Resuming to the last-seen
-        // panel across sessions is confusing — a user re-opens the book and Panel View lands
-        // them mid-page in a way that reads as "starts on the right" (or wherever they
-        // happened to close). Panel index still walks correctly during the reading session
-        // (nextPanel / previousPanel) and page turns reset it to 0.
-        _currentPanelIndex.value = 0
-
-        panelBook = panelEngine.forBook(
-            bookId = bookId,
-            imageBytes = { pageIndex -> source.imageBytes(pageIndex) },
-        )
-        _state.value = CbzReaderState.Ready(
-            title = title,
-            pageCount = pageCount,
-            imageSource = source,
-        )
-        val payload = lastPosition?.takeIf { it.isNotEmpty() }?.let {
-            SessionPayload(ebookLocation = it, ebookProgress = 0f)
-        } ?: SessionPayload("", 0f)
-        syncSession.sync(payload)
-        // Prefetch panels for the resume page and the next two.
-        onCurrentPageChanged(resumeIndex)
-    }
-
-    private fun loadStreaming(result: CbzOpenResult.Streaming, item: LibraryItem) {
-        if (result.pageCount == 0) {
-            _state.value = CbzReaderState.Error("Comic has no pages")
-            return
-        }
-        val resumeIndex = result.lastPosition
-            ?.let { parsePageIndex(it) }
-            ?.coerceIn(0, result.pageCount - 1)
-            ?: 0
-        _currentPage.value = resumeIndex
-        lastSavedPage = resumeIndex
-        _currentPanelIndex.value = 0
-
-        // The streaming source (built by the repository) read-aheads the next pages so a
-        // page turn is served from its byte cache instead of a cold synchronous download.
-        val networkSource = result.imageSource
-        currentSource = networkSource
-        panelBook = panelEngine.forBook(
-            bookId = bookId,
-            imageBytes = { pageIndex -> networkSource.imageBytes(pageIndex) },
-        )
-        _state.value = CbzReaderState.Ready(
-            title = item.title,
-            pageCount = result.pageCount,
-            imageSource = networkSource,
-            thumbnailSource = result.thumbnailSource,
-        )
-
-        val payload = result.lastPosition?.takeIf { it.isNotEmpty() }?.let {
-            SessionPayload(ebookLocation = it, ebookProgress = 0f)
-        } ?: SessionPayload("", 0f)
-        syncSession.sync(payload)
-        onCurrentPageChanged(resumeIndex)
-
-        // Background: download the full file and swap to the local archive once ready.
-        viewModelScope.launch {
-            val local = cbzRepository.awaitCachedSource(item) ?: return@launch
-            swapToLocalArchive(local, item.title)
-        }
-    }
-
-    private fun swapToLocalArchive(local: CbzLocalSource, title: String) {
-        val newSource = local.imageSource
-        val actualPageCount = local.pageCount
-        if (archiveClosed) {
-            newSource.close()
-            return
-        }
-        val current = _state.value as? CbzReaderState.Ready
-        if (current == null) {
-            newSource.close()
-            return
-        }
-        // Release the streaming source (cancels its read-ahead scope) + its thumbnail source
-        // before swapping to the local archive.
-        currentSource?.close()
-        current.thumbnailSource?.close()
-        currentSource = newSource
-        val swapPageCount = actualPageCount.takeIf { it > 0 } ?: current.pageCount
-        _railSegments.value = buildCbzRailSegments(local.bookmarks, swapPageCount)
-        _pageCount.value = swapPageCount
-        panelBook = panelEngine.forBook(
-            bookId = bookId,
-            imageBytes = { pageIndex -> newSource.imageBytes(pageIndex) },
-        )
-        // If the server-reported page count differed from the actual archive, clamp current page.
-        clampPageForSwap(_currentPage.value, actualPageCount)?.let { clamped ->
-            _currentPage.value = clamped
-            lastSavedPage = clamped
-        }
-        _state.value = computeArchiveSwapState(current, actualPageCount, newSource)
-        onCurrentPageChanged(_currentPage.value)
-    }
-
-    private fun parsePageIndex(json: String): Int? = try {
-        if (json.startsWith("{")) {
-            JSONObject(json).optJSONObject("locations")
-                ?.let { if (it.has("position")) it.getInt("position") - 1 else null }
-        } else {
-            json.trim().toIntOrNull()
-        }
-    } catch (_: Exception) {
-        null
-    }
-
-    fun nextPage() {
-        val ready = _state.value as? CbzReaderState.Ready ?: return
-        val next = (_currentPage.value + 1).coerceAtMost(ready.pageCount - 1)
-        if (next != _currentPage.value) {
-            gotoPage(next)
-            savePosition(next, ready.pageCount)
-        }
-    }
-
-    fun previousPage() {
-        val next = (_currentPage.value - 1).coerceAtLeast(0)
-        if (next != _currentPage.value) {
-            gotoPage(next)
-            val ready = _state.value as? CbzReaderState.Ready ?: return
-            savePosition(next, ready.pageCount)
-        }
-    }
-
-    fun jumpToPage(index: Int) {
-        val ready = _state.value as? CbzReaderState.Ready ?: return
-        val clamped = index.coerceIn(0, ready.pageCount - 1)
-        if (clamped != _currentPage.value) {
-            gotoPage(clamped)
-            savePosition(clamped, ready.pageCount)
-        }
-    }
-
-    /**
-     * Central page-change primitive. Clears `_currentPagePanels` (so Panel View doesn't render
-     * the new page's bitmap through the previous page's panel geometry) and resets the panel
-     * index before kicking off the async resolve.
-     */
-    private fun gotoPage(newPage: Int) {
-        _currentPage.value = newPage
-        _currentPanelIndex.value = 0
-        _currentPagePanels.value = null
-        onCurrentPageChanged(newPage)
-    }
-
-    // --- Panel View (ADR 0055) ---
-
-    /**
-     * Merges the given [patch] into the current per-book overrides and persists.
-     * Only non-null fields in [patch] replace the corresponding field in the stored overrides.
-     */
-    fun updateComicFormatting(patch: BookComicFormattingOverrides) {
-        val merged = mergeComicFormattingOverrides(_bookComicOverrides.value, patch)
-        _bookComicOverrides.value = merged
-        viewModelScope.launch { bookComicFormattingPreferencesStore.save(bookId, merged) }
-    }
-
-    /** Clears all per-book Comic formatting overrides, falling back to global defaults. */
-    fun resetComicFormattingToDefaults() {
-        _bookComicOverrides.value = BookComicFormattingOverrides()
-        viewModelScope.launch { bookComicFormattingPreferencesStore.reset(bookId) }
-    }
-
-    /**
-     * Advance to the next panel on the current page, or to the first panel of the next page when
-     * the current panel is the last one on the page. When Panel View is off this behaves as
-     * [nextPage] so tap-right / swipe / vol-down continue to work as they always did.
-     */
-    fun nextPanel() {
-        if (!panelViewOn.value) return nextPage()
-        // Seam hold: raw panels are known but effectivePanels is withheld while the SMART_SPLIT
-        // seam decode runs. Falling through to nextPage() here would skip a page the user never
-        // saw a single panel of — swallow the tap instead; the hold lasts tens of milliseconds.
-        if (effectivePanels.value == null && _currentPagePanels.value != null) return
-        val panels = effectivePanels.value?.panels
-        if (panels.isNullOrEmpty() || effectivePanels.value?.isFallback == true) return nextPage()
-        val nextIndex = _currentPanelIndex.value + 1
-        if (nextIndex < panels.size) {
-            _currentPanelIndex.value = nextIndex
-        } else {
-            nextPage()
-        }
-    }
-
-    fun previousPanel() {
-        if (!panelViewOn.value) return previousPage()
-        // Same seam-hold swallow as nextPanel — don't turn a withheld emission into a page jump.
-        if (effectivePanels.value == null && _currentPagePanels.value != null) return
-        val panels = effectivePanels.value?.panels
-        if (panels.isNullOrEmpty() || effectivePanels.value?.isFallback == true) return previousPage()
-        val prevIndex = _currentPanelIndex.value - 1
-        if (prevIndex >= 0) {
-            _currentPanelIndex.value = prevIndex
-        } else {
-            // Cross-page backwards: land on the last panel of the previous page.
-            // gotoPage() cancels the current resolve; we start a new one on Dispatchers.Default
-            // (resolvePage does zip I/O + bitmap decode + detection — never run on Main) and
-            // set the landing index to the last panel once panels are known.
-            val newPage = (_currentPage.value - 1).coerceAtLeast(0)
-            if (newPage == _currentPage.value) return
-            gotoPage(newPage)
-            // Track this landing-index job separately: overwriting panelResolveJob here would
-            // orphan the resolve gotoPage just launched (whose sampler/prefetch work must stay
-            // cancellable by the NEXT page change).
-            panelLandingJob?.cancel()
-            panelLandingJob = viewModelScope.launch {
-                val book = panelBook ?: return@launch
-                val pagePanels = withContext(Dispatchers.Default) {
-                    runCatching { book.resolvePage(newPage) }.getOrNull()
-                } ?: return@launch
-                if (_currentPage.value != newPage) return@launch  // user navigated away mid-resolve
-                _currentPagePanels.value = pagePanels
-                val formatting = effectiveComicFormatting.value
-                val (vpW, vpH) = _viewportSize.value
-                val effectiveCount = if (
-                    formatting.panelViewOn &&
-                    (formatting.panelOverflow == PanelOverflowBehavior.SPLIT ||
-                        formatting.panelOverflow == PanelOverflowBehavior.SMART_SPLIT) &&
-                    !pagePanels.isFallback && vpW > 0 && vpH > 0
-                ) {
-                    // Pass the actual behavior; SMART_SPLIT with no sampler falls back to center
-                    // split, which produces the same count as SPLIT — but using the real behavior
-                    // stays correct if counts ever diverge.
-                    PanelOverflowTransform.applyOverflow(
-                        pagePanels.panels, pagePanels.imageWidth, pagePanels.imageHeight,
-                        vpW, vpH, formatting.panelOverflow,
-                    ).size
-                } else {
-                    pagePanels.panels.size
-                }
-                _currentPanelIndex.value = (effectiveCount - 1).coerceAtLeast(0)
-            }
-        }
-    }
-
-    /**
-     * Escape hatch for a mis-ordered page (long-press peek → "skip guided panels on this page"):
-     * jumps directly to the first panel of the next page.
-     */
-    fun skipGuidedPanelsOnPage() = nextPage()
-
-    private fun onCurrentPageChanged(pageIndex: Int) {
-        val book = panelBook ?: return
-        val ready = _state.value as? CbzReaderState.Ready
-        // Cancel any prior in-flight resolve/prefetch so a stale coroutine can't clobber
-        // `_currentPagePanels` with an older page's result under rapid navigation.
-        panelResolveJob?.cancel()
-        // Pending only when a seam decode will actually follow (SMART_SPLIT); otherwise mark
-        // resolved immediately so effectivePanels never holds waiting for a decode that will
-        // never run.
-        _energySampler.value = if (
-            effectiveComicFormatting.value.panelOverflow == PanelOverflowBehavior.SMART_SPLIT
-        ) {
-            EnergySamplerState.Pending
-        } else {
-            EnergySamplerState.Resolved(null)
-        }
-        panelResolveJob = viewModelScope.launch {
-            val resolved = withContext(Dispatchers.Default) {
-                runCatching { book.resolvePage(pageIndex) }.getOrNull()
-            }
-            if (archiveClosed || _currentPage.value != pageIndex) return@launch
-            if (resolved == null) {
-                // Detection failed (corrupt/unreadable page). Emit a synthetic fallback so the
-                // viewer shows the page fit-whole instead of an endless spinner — the image
-                // gate holds while pagePanels is null, and null would otherwise stay forever.
-                // Resolve the sampler too so a Pending state can't outlive its decode.
-                _currentPagePanels.value = failedDetectionFallback(pageIndex)
-                _energySampler.value = EnergySamplerState.Resolved(null)
-                return@launch
-            }
-            val current = resolved
-            _currentPagePanels.value = current
-            // Resolve the seam sampler whenever this page set Pending — keyed on the STATE, not
-            // a re-read of the formatting: if the user toggles SMART_SPLIT off (or off and back
-            // on) mid-decode, a formatting re-check would strand Pending forever and the seam
-            // hold would spin until the next page turn. A failed decode resolves to a null
-            // sampler (centre-split fallback), never stays Pending.
-            if (_energySampler.value == EnergySamplerState.Pending) {
-                val computed = withContext(Dispatchers.Default) {
-                    computeEnergySampler(pageIndex, current.imageWidth, current.imageHeight)
-                }
-                if (archiveClosed || _currentPage.value != pageIndex) {
-                    // Page moved on while we scanned — this result belongs to a page that is no
-                    // longer current; publishing it would release the new page's hold with a
-                    // stale seam.
-                    return@launch
-                }
-                _energySampler.value = EnergySamplerState.Resolved(computed)
-            }
-            // Prefetch the next two pages.
-            withContext(Dispatchers.Default) {
-                for (offset in 1..2) {
-                    if (archiveClosed) break
-                    val target = pageIndex + offset
-                    if (ready != null && target >= ready.pageCount) break
-                    runCatching { book.resolvePage(target) }
-                }
-            }
-        }
-    }
-
-    // --- Position sync ---
-
-    private fun savePosition(pageIndex: Int, pageCount: Int) {
-        if (pageIndex == lastSavedPage) return
-        lastSavedPage = pageIndex
-        val progression = if (pageCount > 0) (pageIndex + 1).toDouble() / pageCount.toDouble() else 0.0
-        val locatorJson = JSONObject()
-            .put("href", "cbz://page")
-            .put("type", "application/vnd.comicbook+zip")
-            .put(
-                "locations", JSONObject()
-                    .put("position", pageIndex + 1)
-                    .put("totalProgression", progression),
-            )
-            .toString()
-        viewModelScope.launch {
-            cbzRepository.saveReadingPosition(itemId, locatorJson)
-            val progressFraction = if (pageCount > 1) pageIndex.toFloat() / (pageCount - 1).toFloat() else 1f
-            updateReadingProgressUseCase(itemId, progressFraction)
-        }
-    }
-
-    fun setKeepScreenOn(value: Boolean) {
-        viewModelScope.launch { wakeLockPreferencesStore.setKeepScreenOn(value) }
-    }
-
-    fun setVolumeKeyNavigationEnabled(value: Boolean) {
-        viewModelScope.launch { volumeKeyDispatcher.setVolumeKeyNavigationEnabled(value) }
-    }
-
-    fun setInvertVolumeKeys(value: Boolean) {
-        viewModelScope.launch { volumeKeyDispatcher.setInvertVolumeKeys(value) }
-    }
-
-    fun onReaderResumed() {
-        readerStateHolder.isReaderActive = true
-        closeSyncDone = false
-    }
-
-    fun onReaderClosed() {
-        readerStateHolder.isReaderActive = false
-        readerStateHolder.isPanelOpen = false
-        if (closeSyncDone) return
-        closeSyncDone = true
-        val ready = _state.value as? CbzReaderState.Ready ?: return
-        savePosition(_currentPage.value, ready.pageCount)
-        val page = _currentPage.value + 1
-        val progression = if (ready.pageCount > 0) page.toDouble() / ready.pageCount.toDouble() else 0.0
-        val locatorJson = JSONObject()
-            .put("href", "cbz://page")
-            .put("type", "application/vnd.comicbook+zip")
-            .put("locations", JSONObject().put("position", page).put("totalProgression", progression))
-            .toString()
-        syncSession.sync(SessionPayload(ebookLocation = locatorJson, ebookProgress = progression.toFloat()))
-    }
-
-    override fun onCleared() {
-        super.onCleared()
-        // Flag the archive closed BEFORE closing so any in-flight background prefetch skips
-        // its store.save on ZipException — otherwise a 1x1 Fallback would be persisted and
-        // served on the next open.
-        archiveClosed = true
-        panelResolveJob?.cancel()
-        currentSource?.close()
-        currentSource = null
-    }
-
-    /**
-     * Returns the SMART_SPLIT seam-energy sampler for [pageIndex], or null when the page can't be
-     * decoded. The android.graphics decode now lives behind [ColorPageDecoder] (Android:
-     * BitmapFactory; iOS: CoreGraphics) and the pixel-gradient math in
-     * [comicSeamEnergySampler] — both shared in commonMain — so there is no `Bitmap` to hold or
-     * recycle: the decoder returns a heap-allocated colour image the sampler closes over, which the
-     * GC reclaims once the page changes.
-     */
-    private fun computeEnergySampler(
-        pageIndex: Int,
-        imageWidth: Int,
-        imageHeight: Int,
-    ): ((panel: PanelRegion, splitHorizontally: Boolean) -> FloatArray)? {
-        val source = (_state.value as? CbzReaderState.Ready)?.imageSource ?: return null
-        val bytes = runCatching { source.imageBytes(pageIndex) }.getOrNull() ?: return null
-        val image = colorPageDecoder.decode(bytes, targetLongEdge = 300) ?: return null
-        return comicSeamEnergySampler(image, imageWidth, imageHeight)
-    }
-}
-
-/**
- * Computes the new [CbzReaderState.Ready] after a streaming→archive swap. Extracted for unit
- * testing — the ViewModel extends AndroidViewModel and can't be instantiated in JVM tests.
- */
-internal fun computeArchiveSwapState(
-    current: CbzReaderState.Ready,
-    actualPageCount: Int,
-    newSource: ComicPageSource,
-): CbzReaderState.Ready {
-    val effectivePageCount = if (actualPageCount > 0) actualPageCount else current.pageCount
-    return current.copy(
-        pageCount = effectivePageCount,
-        imageSource = newSource,
-        thumbnailSource = null,
-    )
-}
-
-internal fun mergeComicFormattingOverrides(
-    current: BookComicFormattingOverrides,
-    patch: BookComicFormattingOverrides,
-): BookComicFormattingOverrides = current.copy(
-    backgroundTheme = patch.backgroundTheme ?: current.backgroundTheme,
-    panelViewOn = patch.panelViewOn ?: current.panelViewOn,
-    panelOverflow = patch.panelOverflow ?: current.panelOverflow,
-    panelAnimationSpeedMs = patch.panelAnimationSpeedMs ?: current.panelAnimationSpeedMs,
-    showChapterMap = patch.showChapterMap ?: current.showChapterMap,
-    showPageProgress = patch.showPageProgress ?: current.showPageProgress,
-)
-
-/**
- * Returns the clamped page index when an archive swap reveals the real page count is smaller
- * than what the server reported during streaming, or null if no clamping is needed.
- */
-internal fun clampPageForSwap(currentPage: Int, actualPageCount: Int): Int? =
-    if (actualPageCount > 0 && currentPage >= actualPageCount) actualPageCount - 1 else null
-
-/**
- * Synthetic PagePanels for a page whose panel detection failed (corrupt/unreadable page).
- * isFallback routes the viewer to the fit-whole path, whose transform is Identity — the fake
- * 1×1 dimensions are never used for camera math.
- */
-internal fun failedDetectionFallback(pageIndex: Int): PagePanels = PagePanels(
-    pageIndex = pageIndex,
-    imageWidth = 1,
-    imageHeight = 1,
-    panels = listOf(PanelRegion(0, 0, 1, 1)),
-    source = PanelSource.Fallback,
-)
-
-/** SMART_SPLIT seam-sampler resolution state for the current page. */
-internal sealed interface EnergySamplerState {
-    /** Page changed with SMART_SPLIT active; the low-res seam decode hasn't finished yet. */
-    data object Pending : EnergySamplerState
-
-    /** Decode finished (or not needed). [sampler] is null when the page couldn't be decoded. */
-    data class Resolved(
-        val sampler: ((panel: PanelRegion, splitHorizontally: Boolean) -> FloatArray)?,
-    ) : EnergySamplerState
-}
-
-/**
- * True when effectivePanels must withhold its result because the SMART_SPLIT seam decode is
- * still in flight AND at least one panel on the page would actually be split. Emitting an
- * interim centre-split would render the page and then visibly re-split (a focus jump) when
- * the real seam arrives; holding keeps the screen in its loading state until the seam is known.
- */
-internal fun shouldHoldForSmartSeam(
-    overflow: PanelOverflowBehavior,
-    samplerPending: Boolean,
-    pagePanels: PagePanels,
-    viewportW: Int,
-    viewportH: Int,
-): Boolean = overflow == PanelOverflowBehavior.SMART_SPLIT && samplerPending &&
-    // Probe with the transform itself (SPLIT = centre seam, same split decision as SMART_SPLIT)
-    // rather than a separate predicate, so "would this page split?" can never drift from what
-    // applyOverflow actually does when the hold releases.
-    PanelOverflowTransform.applyOverflow(
-        pagePanels.panels, pagePanels.imageWidth, pagePanels.imageHeight,
-        viewportW, viewportH, PanelOverflowBehavior.SPLIT,
-    ).size != pagePanels.panels.size
