+++ feature/reader/build.gradle.kts
@@ -13,6 +13,7 @@ kotlin {
             api(project(":core:models"))
             implementation(libs.kotlinx.coroutines.core)
             implementation(libs.kotlinx.serialization.json)
+            implementation(libs.androidx.lifecycle.viewmodel)
         }
         commonTest.dependencies {
             implementation(kotlin("test"))
